restored to previous workings
