+++ packages/hippodrome-ui/src/pages/LockAndMint.tsx
@@ -1,24 +1,26 @@
 import React from "react"
 import { VStack, Spinner } from "@chakra-ui/react"
 import { useParams } from "react-router-dom"
+import { useQuery } from "../hooks/useQuery"
 import { useLockAndMint } from "../hooks/useRen"
 import { KnownInputChains } from "../models/ren"
 import AwaitingDeposit from "../components/transaction/AwaitingDeposit"
 import AwaitingMint from "../components/transaction/AwaitingMint"
 
 const Transaction: React.FC = () => {
-  const { asset, to, nonce, outputToken } = useParams<{
+  const { asset, to, nonce } = useParams<{
     asset: KnownInputChains
     to: string
     nonce: string
-    outputToken: string
   }>()
+  const query = useQuery()
+  const swap = query.get("swap")
 
   const { lockAndMint, deposits, loading } = useLockAndMint({
     lockNetwork: asset,
     to,
     nonce: parseInt(nonce),
-    outputToken,
+    outputToken: (swap || ''),
   })
 
   if (loading) {
